Completed migration of Linked List
