+++ docs/linkedlist/README.md
@@ -1,13 +1,13 @@
 ## Linked List
 
-1.  [Reverse Linked List](Easy/ReverseLinkedList)
-2.  [Middle Node of Linked List](Easy/MiddleNodeOfLinkedList)
-3.  [Check if Linked List Has a Cycle](Easy/CheckIfLinkedListHasACycle)
-4.  [Insert in Sorted Linked List](Easy/InsertInSortedLinkedList)
-5.  [Merge Two Sorted Linked Lists](Easy/MergeTwoSortedLinkedLists)
-6.  [Reorder Linked List](Hard/ReOrderLinkedList)
-7.  [Partition Linked List](Medium/PartitionLinkedList)
-8.  [Add Two Numbers](Medium/AddTwoNumbers)
-9.  [Merge Sort Linked List](Medium/MergeSortLinkedList)
-10. [Check if Linked List is Palindrome](Medium/CheckIfLinkedListIsPalindrome)
-11. [Cycle Node in Linked List](Medium/CycleNodeInLinkedList)
+1.  [Reverse Linked List](ReverseLinkedList.md)
+2.  [Middle Node of Linked List](MiddleNodeOfLinkedList.md)
+3.  [Check if Linked List Has a Cycle](CheckIfLinkedListHasACycle.md)
+4.  [Insert in Sorted Linked List](InsertInSortedLinkedList.md)
+5.  [Merge Two Sorted Linked Lists](MergeTwoSortedLinkedLists.md)
+6.  [Reorder Linked List](ReOrderLinkedList.md)
+7.  [Partition Linked List](PartitionLinkedList.md)
+8.  [Add Two Numbers](AddTwoNumbers.md)
+9.  [Merge Sort Linked List](MergeSortLinkedList.md)
+10. [Check if Linked List is Palindrome](CheckIfLinkedListIsPalindrome.md)
+11. [Cycle Node in Linked List](CycleNodeInLinkedList.md)
